implement resizing of images prior to prediction
the input tensor for the frozen graph has maximum dimensions that depend
on how the model was exported. Exporting with higher dimensions works but
also increases the time needed for computation.
This change makes sure that images are resized to a maximum of 521x521 pixels
while keeping the aspect ratio. The resize ratio is stored and then used to scale
the image up again prior to trait extraction. The Upscaling uses nearest neighbour
interpolation, and thus might introduce a bit or error.
traits.py has been refactored to not depend on PIL anymore.