refactor rescaling of large input images
previously there was a glitch in creating the overlays when input images were too large.

Because the model currently only accepts input shapes < 602x602, larger
images need to be downscaled to fit.
This is done before creating the records, and was previously reverted
after the prediction so that traits were calculated based on the full
size image and an (nearest-neighbour)-upscaled mask.

We now store the scale factor for each image and use downscaled
images for trait extraction and diagnostic images. Pixel counts
for each class are calculated based on the scale factor,
and regionprops are calculated based on a upscaled binary mask.

For color-channel statistics, it is probably safe to ignore this
because pixel values are averaged when downscaling the original.