catch ValueError when calculation region props
since we're doing per-class regionprops, it seems that the inertia
tensor eigenvalue l2 can for some reason have slightly negative
values close to zero. The minor axis length is 4 * sqrt(l2)
which is not defined and thus causes a ValueError to be raised.

Now the result is set to NA if the error occurs.