relabel custom masks before adding colormap
reverts back to the old and simple method to map labels to colors based on the index in
the colormap array.
This approach however requires relabeling of custom masks to be in the range 0..3 because the colormap
only contains as many elements as there are segmentation classes

this change is needed because label2rgb does not reliably assign the same color to the same segmentation class
across images, for example when not all classes are present in a given mask.